+++ api/v2api/doc.go
@@ -1,0 +1,5 @@
+// Package v2api provides an interface for interacting with the v2 full node APIs
+// within the Filecoin network.
+//
+// This package is experimental and the API may change without notice.
+package v2api